outlined actions and their effects in current code design (thinking of changing how the effects work)

+++ src/genetic_code.rs
@@ -2,7 +2,7 @@ pub use crate::command::Command;
 pub use crate::color::Color;
 
 /* parameters for manipulation with genetic code, later can be moved on JS side */
-const GENETIC_CODE_SIZE:usize = 256;
+const GENETIC_CODE_SIZE:i32 = 256;
 
 pub struct GeneticCode {
     genetic_tape:Vec<Command>,
@@ -25,8 +25,8 @@ impl GeneticCode {
         self.rotate_tape_by(1);
         self.genetic_tape[self.current_genetic_tape_ptr]
     }
-    pub fn rotate_tape_by(&mut self, val:usize){
-        self.current_genetic_tape_ptr = (self.current_genetic_tape_ptr+val)%GENETIC_CODE_SIZE;
+    pub fn rotate_tape_by(&mut self, val:i32){
+        self.current_genetic_tape_ptr = ((self.current_genetic_tape_ptr as i32 +val)%GENETIC_CODE_SIZE) as usize;
     }
 }
